+++ src/data/protocols/db/account/add-account-repository.ts
@@ -1,0 +1,6 @@
+import { AddAccountModel } from '../../../../domain/usecases/add-account'
+import { AccountModel } from '../../../../domain/models/account'
+
+export interface AddAccountRepository {
+  add (accountData: AddAccountModel): Promise<AccountModel>
+}
