+++ src/infra/db/mongodb/log/log-ongo-repository.spec.ts
@@ -1,6 +1,6 @@
 import { Collection } from 'mongodb'
 import { MongoHelper } from '../helpers/mongo-helper'
-import { LogMongoRepository } from './log'
+import { LogMongoRepository } from './log-ongo-repository'
 
 const makeSut = (): LogMongoRepository => {
   return new LogMongoRepository()
