+++ src/infra/db/mongodb/log/log-ongo-repository.ts
@@ -1,4 +1,4 @@
-import { LogErrorRepository } from '../../../../data/protocols/db/log-error-repository'
+import { LogErrorRepository } from '../../../../data/protocols/db/log/log-error-repository'
 import { MongoHelper } from '../helpers/mongo-helper'
 
 export class LogMongoRepository implements LogErrorRepository {
