+++ src/presentation/controllers/signup/signup-controller.ts
@@ -1,5 +1,5 @@
 
-import { HttpRequest, HttpResponse, Controller, AddAccount, Validation } from './signup-protocols'
+import { HttpRequest, HttpResponse, Controller, AddAccount, Validation } from './signup-controller-protocols'
 import { badRequest, serverError, ok } from '../../helpers/http/http-helper'
 
 export class SignUpController implements Controller {
